Select only columns required by the query
This significantly improves performance, as the database doesn't have to send so much data to the server, and the server doesn't have to create so many objects when it receives the data.

Mostly, we include every column that's queried in at least one edition. There's special handling for details, as that is usually needed in at least one document, and can be expensive to load for linked documents. We pass whether or not the selection needs details along with the link type path to look up, and respond with null using a CASE WHEN if the edition at this path doesn't need details.