Prepare statements / cache prepared statements
This massively reduces the amount of work Sequel has to do on each request (preparing the big SQL query etc.), and should help the database with parsing time (and maybe planning time).

Unfortunately, I've had to do some pretty ugly things with caching to work around a worry I have with unbounded growth of the prepared statements. Remembering that this is a proof of concept, and that speed is the primary goal... this seems acceptable for now. We might have to get in touch with the maintainers of Sequel to ask their advice on a better approach.