Nullify columns not needed by the query
Previously we were selecting all the columns needed by at least one field in the entire query. This performed well enough, but it meant that we had to create whole new prepared statements for every combination of columns needed.

Instead, we pass which columns each layer of the query needs
as input, and nullify everything else (using case / when).

This means we don't have to load unnecessary data over the network, or unnecessarily allocate ruby objects for values we don't need.