Enable CSRF protection for requests without bearer tokens
If a request does have a bearer token, then we can assume we don't need the cookie for authentication and nullify the session as a CSRF protection.

If the request doesn't have a bearer token, then assume it's the graphiql client in this app, and use the session cookie for authentication and perform all the usual CSRF checks.