Add a content store shim endpoint
This should behave similarly to the real content store in most cases,
assuming we've correctly implemented the graphql schema and the queries
we generated are correct.

Some schemas may perform badly (e.g. those that expand HTML fields like
body in high-cardinality links). We should be able to look at tuning
these, as it's rare that the frontend actually needs all the stuff in
the content item.