Support graphql aliases on links_of_type
In some queries, we have:

parent_taxons: links_of_type(type: "parent_taxon")
child_taxons: links_of_type(type: "parent_taxon", reverse: true)

Which means that link type isn't enough to disambiguate one from the
other. The alias has to be unique, so we can use that instead of the
link type as the path key.

This means we have to mess around with ast_nodes a bit, which is ugly.
But this whole implementation is ugly so ¯\_(ツ)_/¯